Added mock functions and reformatted .sol files

+++ brownie/contracts/StandardAuction.sol
@@ -7,35 +7,39 @@ import "@chainlink/contracts/src/v0.8/VRFConsumerBase.sol";
 import "@chainlink/contracts/src/v0.8/interfaces/KeeperCompatibleInterface.sol";
 import "./MintAuctionBase.sol";
 
-// Auctions are set by a seller and automatically resolved via chainlink keepers. Atm this 
+// Auctions are set by a seller and automatically resolved via chainlink keepers. Atm this
 // resolution is done on a first come first serve basis. It would take a massive amount of auctions
 // needing resolution to make this untollerable. In the future I hope to make this a little smoother
 // by notifying checkupkeep work needs to be done and on who, rather then having to check for who.
 // for more detail look in AuctionBase.sol
 // for more detail look in AuctionBase.sol
 
-contract StandardAuction is MintAuction, ERC1155Holder, VRFConsumerBase, KeeperCompatibleInterface {
+contract StandardAuction is
+    MintAuction,
+    ERC1155Holder,
+    VRFConsumerBase,
+    KeeperCompatibleInterface
+{
     bytes32 public x_keyHash;
     uint256 public x_chainlinkFee;
+
     constructor(
-        uint256 _networkCut,  // a fee that goes towards sustaining this autonomas platform.
-        address _vrfCoordinator,  
-        address _linkToken,     // link token address
-        bytes32 _keyHash,       
-        uint256 _chainlinkFee   // 0.01
-        ) VRFConsumerBase(_vrfCoordinator, _linkToken) { 
+        uint256 _networkCut, // a fee that goes towards sustaining this autonomas platform.
+        address _vrfCoordinator,
+        address _linkToken, // link token address
+        bytes32 _keyHash,
+        uint256 _chainlinkFee // 0.01
+    ) VRFConsumerBase(_vrfCoordinator, _linkToken) {
         require(_networkCut <= 10000);
         networkFee = _networkCut;
         x_chainlinkFee = _chainlinkFee;
         x_keyHash = _keyHash;
     }
-    // testing function 
+
+    // testing function
     function withdrawlBalance(address _nft, uint256 _tokenId) external payable {
         ERC1155 nft = ERC1155(_nft);
-        require(
-            _ownerOf(_nft, _tokenId) ||
-            msg.sender == address(nft)
-        );
+        require(_ownerOf(_nft, _tokenId) || msg.sender == address(nft));
         payable(msg.sender).transfer(address(this).balance);
     }
 
@@ -49,9 +53,8 @@ contract StandardAuction is MintAuction, ERC1155Holder, VRFConsumerBase, KeeperC
         uint256 _startedAt,
         uint128 _highestBid,
         address _highestBidder
-    )public payable
-    {
-        require(_ownerOf(_nft,_tokenId));
+    ) public payable {
+        require(_ownerOf(_nft, _tokenId));
         _escrow(msg.sender, _nft, _tokenId);
         Auction memory auction = Auction(
             payable(_nft),
@@ -65,23 +68,26 @@ contract StandardAuction is MintAuction, ERC1155Holder, VRFConsumerBase, KeeperC
         );
         _addAuction(auction);
     }
+
     // make a bid
-    function bid(address _nft, uint256 _tokenId) public payable{
+    function bid(address _nft, uint256 _tokenId) public payable {
         _bid(_nft, _tokenId);
     }
-    // withdraw old bids.     
-    function withdrawBid() public payable{
+
+    // withdraw old bids.
+    function withdrawBid() public payable {
         _withdrawReturn();
     }
 
     // checks status / prob to be removed as we show this in getAuction
-    function isOn(address _nft, uint256 _tokenId) public view returns(bool){
+    function isOn(address _nft, uint256 _tokenId) public view returns (bool) {
         bool x = _onAuction(_nft, _tokenId);
         return x;
     }
+
     // to be removed
-    function onEndAuction(address _nft, uint256 _tokenId) public payable{
-        _End(_nft,_tokenId);
+    function onEndAuction(address _nft, uint256 _tokenId) public payable {
+        _End(_nft, _tokenId);
     }
 
     // cancel auction
@@ -91,78 +97,86 @@ contract StandardAuction is MintAuction, ERC1155Holder, VRFConsumerBase, KeeperC
         require(_onAuction(_nft, _tokenId));
         _cancelAuction(auction.seller, _nft, _tokenId);
     }
+
     // get all auctions
-    function getAuctions() public view returns(Auction[] memory){
+    function getAuctions() public view returns (Auction[] memory) {
         Auction[] memory ref;
-        for(int i = 0; i < int(autoResolveQ.length); i++){
-            ref[uint(i)] = autoResolveQ[uint(i)];
+        for (int256 i = 0; i < int256(autoResolveQ.length); i++) {
+            ref[uint256(i)] = autoResolveQ[uint256(i)];
         }
         return ref;
     }
+
     // get an auction
     function getAuction(address _nft, uint256 _tokenId)
-     public view returns(
-        address,
-        address,
-        uint256,
-        uint256,
-        uint256,
-        bool
+        public
+        view
+        returns (
+            address,
+            address,
+            uint256,
+            uint256,
+            uint256,
+            bool
         )
-        {
-            Auction storage auction = tokenAddressAndIdToAuction[_nft][_tokenId];
-            require(_onAuction(_nft, _tokenId));
-                return  (
-                    address(auction.seller),
-                    address(auction.nft),
-                    uint256(auction.tokenId),
-                    uint256(auction.highestBid),
-                    uint256(auction.duration),
-                    bool(_onAuction(_nft, _tokenId))
-                );
-        }
+    {
+        Auction storage auction = tokenAddressAndIdToAuction[_nft][_tokenId];
+        require(_onAuction(_nft, _tokenId));
+        return (
+            address(auction.seller),
+            address(auction.nft),
+            uint256(auction.tokenId),
+            uint256(auction.highestBid),
+            uint256(auction.duration),
+            bool(_onAuction(_nft, _tokenId))
+        );
+    }
+
     // Im thinking that we use this to track ids still but in a more secondary way
     // using a memory reference to interface with autoResolveQ and stash indexs or send particular credentials ( NFT, tokenId )
-    //uint[] public indexRef; 
-    function checkUpkeep(bytes memory /*checkData*/)
+    //uint[] public indexRef;
+    function checkUpkeep(
+        bytes memory /*checkData*/
+    )
         public
         view
         override
         returns (bool upKeepNeeded, bytes memory performData)
     {
         Auction[] memory ref = autoResolveQ;
-        uint[] memory indexRef; 
-        uint count = 0;
+        uint256[] memory indexRef;
+        uint256 count = 0;
         bool go = false;
-        for(int i = 0; i < int(ref.length); i++){
+        for (int256 i = 0; i < int256(ref.length); i++) {
             address nft = ref[uint256(i)].nft;
             uint256 tokenId = ref[uint256(i)].tokenId;
-            if(!_onAuction(nft, tokenId)){
+            if (!_onAuction(nft, tokenId)) {
                 go = true;
-                indexRef[count] = uint(i); 
+                indexRef[count] = uint256(i);
                 count++;
             }
-        } 
+        }
         upKeepNeeded = (go);
         performData = abi.encode(indexRef); // bytes("");
     }
+
     event PerformedUPKEEP();
+
     function performUpkeep(bytes calldata performData) external override {
-        uint[] memory ref = abi.decode(performData, (uint[]));
+        uint256[] memory ref = abi.decode(performData, (uint256[]));
         //require(address(this).balance >= 0, "Not enough ETH");
         (bool upKeepNeeded, ) = checkUpkeep("");
         require(upKeepNeeded, "No upkeep needed");
         //resolve an auction
-        for(int i = 0; i < int(ref.length); i++){
-            uint idx = ref[uint(i)];
+        for (int256 i = 0; i < int256(ref.length); i++) {
+            uint256 idx = ref[uint256(i)];
             Auction storage auc = autoResolveQ[idx]; // mock auction holding the id and address
             _End(auc.nft, auc.tokenId);
-            autoResolveQ[idx] = autoResolveQ[autoResolveQ.length-1];
+            autoResolveQ[idx] = autoResolveQ[autoResolveQ.length - 1];
             autoResolveQ.pop();
-        
         }
         //delete ref;
-        emit PerformedUPKEEP(); 
+        emit PerformedUPKEEP();
     }
 
     // satisfies reqs
