+++ brownie/contracts/test/EnglishAuction.sol
@@ -1,45 +1,103 @@
 // SPDX-License-Identifier: MIT
 pragma solidity ^0.8.7;
 
-interface IERC1155 /* is ERC165 */ {
-    event TransferSingle(address indexed _operator, address indexed _from, address indexed _to, uint256 _id, uint256 _value);
-    event TransferBatch(address indexed _operator, address indexed _from, address indexed _to, uint256[] _ids, uint256[] _values);
-    event ApprovalForAll(address indexed _owner, address indexed _operator, bool _approved);
+/* is ERC165 */
+interface IERC1155 {
+    event TransferSingle(
+        address indexed _operator,
+        address indexed _from,
+        address indexed _to,
+        uint256 _id,
+        uint256 _value
+    );
+    event TransferBatch(
+        address indexed _operator,
+        address indexed _from,
+        address indexed _to,
+        uint256[] _ids,
+        uint256[] _values
+    );
+    event ApprovalForAll(
+        address indexed _owner,
+        address indexed _operator,
+        bool _approved
+    );
     event URI(string _value, uint256 indexed _id);
-    function safeTransferFrom(address _from, address _to, uint256 _id, uint256 _value, bytes calldata _data) external;
-    function safeBatchTransferFrom(address _from, address _to, uint256[] calldata _ids, uint256[] calldata _values, bytes calldata _data) external;
-    function balanceOf(address _owner, uint256 _id) external view returns (uint256);
-    function balanceOfBatch(address[] calldata _owners, uint256[] calldata _ids) external view returns (uint256[] memory);
+
+    function safeTransferFrom(
+        address _from,
+        address _to,
+        uint256 _id,
+        uint256 _value,
+        bytes calldata _data
+    ) external;
+
+    function safeBatchTransferFrom(
+        address _from,
+        address _to,
+        uint256[] calldata _ids,
+        uint256[] calldata _values,
+        bytes calldata _data
+    ) external;
+
+    function balanceOf(address _owner, uint256 _id)
+        external
+        view
+        returns (uint256);
+
+    function balanceOfBatch(address[] calldata _owners, uint256[] calldata _ids)
+        external
+        view
+        returns (uint256[] memory);
+
     function setApprovalForAll(address _operator, bool _approved) external;
-    function isApprovedForAll(address _owner, address _operator) external view returns (bool);
+
+    function isApprovedForAll(address _owner, address _operator)
+        external
+        view
+        returns (bool);
 }
+
 interface ERC1155TokenReceiver {
-    function onERC1155Received(address _operator, address _from, uint256 _id, uint256 _value, bytes calldata _data) external returns(bytes4);
-    function onERC1155BatchReceived(address _operator, address _from, uint256[] calldata _ids, uint256[] calldata _values, bytes calldata _data) external returns(bytes4);
+    function onERC1155Received(
+        address _operator,
+        address _from,
+        uint256 _id,
+        uint256 _value,
+        bytes calldata _data
+    ) external returns (bytes4);
+
+    function onERC1155BatchReceived(
+        address _operator,
+        address _from,
+        uint256[] calldata _ids,
+        uint256[] calldata _values,
+        bytes calldata _data
+    ) external returns (bytes4);
 }
 
 contract EnglishAuction {
     event Start();
-    event Bid(address indexed sender, uint amount);
-    event Withdraw(address indexed bidder, uint amount);
-    event End(address winner, uint amount);
+    event Bid(address indexed sender, uint256 amount);
+    event Withdraw(address indexed bidder, uint256 amount);
+    event End(address winner, uint256 amount);
 
     IERC1155 public nft;
-    uint public nftId;
+    uint256 public nftId;
 
     address payable public seller;
-    uint public endAt;
+    uint256 public endAt;
     bool public started;
     bool public ended;
 
     address public highestBidder;
-    uint public highestBid;
-    mapping(address => uint) public bids;
+    uint256 public highestBid;
+    mapping(address => uint256) public bids;
 
     constructor(
         address _nft,
-        uint _nftId,
-        uint _startingBid
+        uint256 _nftId,
+        uint256 _startingBid
     ) {
         nft = IERC1155(_nft);
         nftId = _nftId;
@@ -50,9 +108,9 @@ contract EnglishAuction {
 
     function start() public {
         require(!started, "started");
-       // require(msg.sender == seller, "not seller");
+        // require(msg.sender == seller, "not seller");
 
-        nft.safeTransferFrom(msg.sender, address(this), nftId, 1, '');
+        nft.safeTransferFrom(msg.sender, address(this), nftId, 1, "");
         started = true;
         endAt = block.timestamp + 7 days;
 
@@ -75,7 +133,7 @@ contract EnglishAuction {
     }
 
     function withdraw() public {
-        uint bal = bids[msg.sender];
+        uint256 bal = bids[msg.sender];
         bids[msg.sender] = 0;
         payable(msg.sender).transfer(bal);
 
@@ -89,10 +147,10 @@ contract EnglishAuction {
 
         ended = true;
         if (highestBidder != address(0)) {
-            nft.safeTransferFrom(address(this), highestBidder, nftId, 1, '');
+            nft.safeTransferFrom(address(this), highestBidder, nftId, 1, "");
             seller.transfer(highestBid);
         } else {
-            nft.safeTransferFrom(seller, address(this), nftId, 1, '');
+            nft.safeTransferFrom(seller, address(this), nftId, 1, "");
         }
 
         emit End(highestBidder, highestBid);
